+++ src/align/bounded/structs/seed.rs
@@ -1,3 +1,6 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
 pub struct Seed {
     pub target_name: String,
     pub target_start: usize,